[REFACTOR] 컨벤션에 맞춰 DTO 형태 수정 #163
