--- src/main/java/com/sports/server/command/leagueteam/dto/LeagueTeamUpdateRequest.java
@@ -1,23 +1,0 @@
-package com.sports.server.command.leagueteam.dto;
-
-import com.sports.server.command.leagueteam.domain.LeagueTeam;
-import com.sports.server.command.leagueteam.domain.LeagueTeamPlayer;
-import java.util.List;
-
-public record LeagueTeamUpdateRequest(
-
-        String name,
-        String logoImageUrl,
-        List<LeagueTeamPlayerRegisterRequest> addPlayers,
-        List<Long> deletedPlayerIds
-
-) {
-    public record LeagueTeamPlayerRegisterRequest(
-            String name,
-            int number
-    ) {
-        public LeagueTeamPlayer toEntity(LeagueTeam leagueTeam) {
-            return new LeagueTeamPlayer(leagueTeam, name, number);
-        }
-    }
-}
